feat: 🚧 Functions for future usage in parsing AST
Created a set of specific use-case functions for future usage while parsing AST. This primarily means that these function are used to parse specific elements of the AST within PS scripts. Currently there usage is not documented externally.

No breaking changes nor issues remediated as a result of this commit.