feat: ✨ Change to output the results in a nicer more structured format.
Changed the pslint `END` block to output the results in a nicer more structured format. Checks the environment variable `[bool]$env:CI` to determine if the execution environment is a part of the CI workflow as that is the overall intent of this module.

No breaking changes nor issues remediated for `$this.commit`.